Use driver.Prepare api for EnsureImageExists
EnsureImageExists is a subset of driver.Prepare as it stands. I don't quite
understand why the check is needed in any case:

for route creation why do we care if the image exists prior to letting it
land? the image could easily be removed from the hub/registry and be in the
same funky state where the route might exist in the api gateway but when a
new server tries to run it it's incapable, so that check doesn't seem a very
good gate (just a sanity check).

if enough agree about ^ we could remove EnsureImageExists on the whole. The
only change this patch makes is such that not only will the image exist, but
it's now more likely that we could actually run a task with that image since
driver.Prepare does some additional checks.