functions: add cache priming
In order to improve the effectiveness of the LRU cache, it primes
the cache with known routes up to the limit of the LRU cache
threshold. It slows the start process, but it pays off by easing
the load on the rest of the structure during thundering herds and
start under heavy load.